Adds personalized config.h
Selects font to be system default monospace font at size 16. Due to
fontsize discrepancies this may change in the future.

Changes tabspaces to 4 spaces, as they should be.

Manually resets colors to be the same as my .Xresources. Thus if I
change my .Xresources, the changes will not be reflected in my terminal.
This was done to avoid the .Xresources patch which appeared to be more
code than it was worth. May reconsider this choice in the future.

Set mousescrollincrement to 4, simply because my preceeding terminal,
urxvt, set this as a default and I enjoy that consistency.